requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
dropbox>=11.36.0
